// mobile/lib/api.d.ts
export interface components {
  schemas: {
// ...
    CandidateDetailOut: {
      /**
       * Candidate Id
       * Format: uuid
       */
      candidate_id: string;
      /** Full Name */
      full_name: string;
      /** Email */
      email: string;
      /** Phone Number */
      phone_number: string | null;
      submission_status: components["schemas"]["SubmissionStatus"];
      /**
       * Submitted At
       * Format: date-time
       */
      submitted_at: string;
      /** Parse Error */
      parse_error: string | null;
      /** Resume Url */
      resume_url: string | null;
      restorable_status: components["schemas"]["SubmissionStatus"];
      /** Form Responses */
      form_responses: components["schemas"]["FormResponseOut"][];
    };
    /** CandidateOut */
    CandidateOut: {
      /**
       * Candidate Id
       * Format: uuid
       */
      candidate_id: string;
      /** Full Name */
      full_name: string;
      /** Email */
      email: string;
      /** Phone Number */
      phone_number: string | null;
      submission_status: components["schemas"]["SubmissionStatus"];
      /**
       * Submitted At
       * Format: date-time
       */
      submitted_at: string;
      /** Parse Error */
      parse_error: string | null;
      /** Resume Url */
      resume_url: string | null;
      restorable_status: components["schemas"]["SubmissionStatus"];
    };
// ...
    Page_CandidateOut_: {
      /** Items */
      items: components["schemas"]["CandidateOut"][];
      /** Total */
      total: number;
      /** Page */
      page: number;
      /** Size */
      size: number;
    };
// ...
    SubmissionStatus:
      | "SUBMITTED"
      | "PARSED"
      | "RANKED"
      | "INVITED"
      | "CONFIRMED"
      | "DECLINED"
      | "REJECTED"
      | "RESCHEDULED"
      | "PARSE_ERROR";
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}